Se corrige distribución de código en distintas paginas y se agregue logo del aplicativo
