+++ frontend/src/components/layout/MainLayout.jsx
@@ -1,0 +1,10 @@
+import { Sidebar } from './Sidebar.jsx';
+
+export function MainLayout({ activePath, children }) {
+  return (
+    <main className="app-shell">
+      <Sidebar activePath={activePath} />
+      {children}
+    </main>
+  );
+}
