+++ frontend/src/components/ui/ActionButton.jsx
@@ -1,0 +1,9 @@
+import { Icon } from './Icon.jsx';
+
+export function ActionButton({ icon, label, onClick }) {
+  return (
+    <button className="action-button" type="button" onClick={onClick} aria-label={label} title={label}>
+      <Icon name={icon} />
+    </button>
+  );
+}
