+++ frontend/src/components/ui/Badge.jsx
@@ -1,0 +1,3 @@
+export function Badge({ className = '', children }) {
+  return <span className={className}>{children}</span>;
+}
