+++ frontend/src/components/ui/Card.jsx
@@ -1,0 +1,4 @@
+export function Card({ className = '', children }) {
+  const cardClass = ['product-form-card', className].filter(Boolean).join(' ');
+  return <section className={cardClass}>{children}</section>;
+}
