+++ frontend/src/components/ui/Input.jsx
@@ -1,0 +1,3 @@
+export function Input({ error, ...props }) {
+  return <input className={error ? 'field-invalid' : ''} {...props} />;
+}
