+++ frontend/src/components/ui/Select.jsx
@@ -1,0 +1,3 @@
+export function Select({ error, children, ...props }) {
+  return <select className={error ? 'field-invalid' : ''} {...props}>{children}</select>;
+}
