+++ frontend/src/components/ui/Table.jsx
@@ -1,0 +1,3 @@
+export function Table({ children, className = '' }) {
+  return <table className={className}>{children}</table>;
+}
