+++ frontend/src/components/ui/Textarea.jsx
@@ -1,0 +1,3 @@
+export function Textarea({ error, ...props }) {
+  return <textarea className={error ? 'field-invalid' : ''} {...props} />;
+}
