+++ frontend/src/data/mockProducts.js
@@ -1,0 +1,58 @@
+export const sampleProducts = [
+  {
+    code: 'PROD-001',
+    name: 'Paracetamol 500mg',
+    description: 'Tabletas',
+    category: 'Medicamentos',
+    presentation: 'Caja x 100',
+    stock: 120,
+    status: 'Activo',
+  },
+  {
+    code: 'PROD-002',
+    name: 'Omeprazol 20mg',
+    description: 'Capsulas',
+    category: 'Medicamentos',
+    presentation: 'Caja x 30',
+    stock: 80,
+    status: 'Activo',
+  },
+  {
+    code: 'PROD-003',
+    name: 'Ibuprofeno 400mg',
+    description: 'Tabletas',
+    category: 'Medicamentos',
+    presentation: 'Caja x 50',
+    stock: 60,
+    status: 'Activo',
+  },
+  {
+    code: 'PROD-004',
+    name: 'Guantes de Nitrilo Talla M',
+    description: 'Caja x 100 unidades',
+    category: 'Insumos Medicos',
+    presentation: 'Caja x 100',
+    stock: 200,
+    status: 'Activo',
+  },
+  {
+    code: 'PROD-005',
+    name: 'Jeringa 5ml',
+    description: 'Esteril, con aguja',
+    category: 'Insumos Medicos',
+    presentation: 'Caja x 100',
+    stock: 150,
+    status: 'Activo',
+  },
+  {
+    code: 'PROD-006',
+    name: 'Alcohol antiseptico 70%',
+    description: 'Frasco 500ml',
+    category: 'Aseo y desinfeccion',
+    presentation: 'Unidad',
+    stock: 0,
+    status: 'Agotado',
+  },
+];
+
+export const productsStorageKey = 'sagi_products';
