+++ frontend/src/data/mockRequests.js
@@ -1,0 +1,236 @@
+export const sampleRequests = [
+  {
+    id: 1,
+    code: 'SOL-2024-0001',
+    date: '2024-05-27T10:30:00',
+    area: 'Farmacia',
+    requestedBy: 'Maria Gomez',
+    status: 'Pendiente',
+    priority: 'Media',
+    observations: 'Reposicion semanal para dispensacion.',
+    createdAt: '2024-05-27T10:30:00',
+    approvedAt: '',
+    approvedBy: '',
+    products: [
+      { product: 'Paracetamol 500mg', productCode: 'PROD-001', quantity: 30, unit: 'Caja', observations: 'Alta rotacion' },
+      { product: 'Omeprazol 20mg', productCode: 'PROD-002', quantity: 12, unit: 'Caja', observations: '' },
+      { product: 'Jeringa 5ml', productCode: 'PROD-005', quantity: 8, unit: 'Caja', observations: 'Con aguja' },
+    ],
+  },
+  {
+    id: 2,
+    code: 'SOL-2024-0002',
+    date: '2024-05-26T16:15:00',
+    area: 'Urgencias',
+    requestedBy: 'Carlos Ramirez',
+    status: 'Aprobada',
+    priority: 'Alta',
+    observations: 'Solicitud para turno nocturno.',
+    createdAt: '2024-05-26T16:15:00',
+    approvedAt: '2024-05-26T17:05:00',
+    approvedBy: 'Admin SAGI',
+    products: [
+      { product: 'Guantes de Nitrilo Talla M', productCode: 'PROD-004', quantity: 20, unit: 'Caja', observations: '' },
+      { product: 'Alcohol antiseptico 70%', productCode: 'PROD-006', quantity: 15, unit: 'Unidad', observations: 'Area triage' },
+      { product: 'Jeringa 5ml', productCode: 'PROD-005', quantity: 10, unit: 'Caja', observations: '' },
+      { product: 'Ibuprofeno 400mg', productCode: 'PROD-003', quantity: 5, unit: 'Caja', observations: '' },
+      { product: 'Paracetamol 500mg', productCode: 'PROD-001', quantity: 8, unit: 'Caja', observations: '' },
+    ],
+  },
+  {
+    id: 3,
+    code: 'SOL-2024-0003',
+    date: '2024-05-25T09:20:00',
+    area: 'Hospitalizacion',
+    requestedBy: 'Laura Torres',
+    status: 'Rechazada',
+    priority: 'Baja',
+    observations: 'Cantidad no justificada para el consumo actual.',
+    createdAt: '2024-05-25T09:20:00',
+    approvedAt: '2024-05-25T11:00:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Omeprazol 20mg', productCode: 'PROD-002', quantity: 50, unit: 'Caja', observations: 'Solicitud sobredimensionada' }],
+  },
+  {
+    id: 4,
+    code: 'SOL-2024-0004',
+    date: '2024-05-24T14:10:00',
+    area: 'Cirugia',
+    requestedBy: 'Andres Molina',
+    status: 'Pendiente',
+    priority: 'Critica',
+    observations: 'Programacion quirurgica de fin de semana.',
+    createdAt: '2024-05-24T14:10:00',
+    approvedAt: '',
+    approvedBy: '',
+    products: [
+      { product: 'Guantes de Nitrilo Talla M', productCode: 'PROD-004', quantity: 40, unit: 'Caja', observations: '' },
+      { product: 'Jeringa 5ml', productCode: 'PROD-005', quantity: 25, unit: 'Caja', observations: '' },
+    ],
+  },
+  {
+    id: 5,
+    code: 'SOL-2024-0005',
+    date: '2024-05-23T08:45:00',
+    area: 'Laboratorio Clinico',
+    requestedBy: 'Natalia Rios',
+    status: 'Aprobada',
+    priority: 'Media',
+    observations: 'Consumo mensual.',
+    createdAt: '2024-05-23T08:45:00',
+    approvedAt: '2024-05-23T10:12:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Alcohol antiseptico 70%', productCode: 'PROD-006', quantity: 18, unit: 'Unidad', observations: '' }],
+  },
+  {
+    id: 6,
+    code: 'SOL-2024-0006',
+    date: '2024-05-22T11:35:00',
+    area: 'Consulta Externa',
+    requestedBy: 'Julian Perez',
+    status: 'Pendiente',
+    priority: 'Baja',
+    observations: 'Reposicion para consultorios.',
+    createdAt: '2024-05-22T11:35:00',
+    approvedAt: '',
+    approvedBy: '',
+    products: [
+      { product: 'Ibuprofeno 400mg', productCode: 'PROD-003', quantity: 4, unit: 'Caja', observations: '' },
+      { product: 'Paracetamol 500mg', productCode: 'PROD-001', quantity: 6, unit: 'Caja', observations: '' },
+    ],
+  },
+  {
+    id: 7,
+    code: 'SOL-2024-0007',
+    date: '2024-05-21T15:00:00',
+    area: 'Urgencias',
+    requestedBy: 'Diana Castro',
+    status: 'Aprobada',
+    priority: 'Critica',
+    observations: 'Contingencia por alta ocupacion.',
+    createdAt: '2024-05-21T15:00:00',
+    approvedAt: '2024-05-21T15:30:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Jeringa 5ml', productCode: 'PROD-005', quantity: 30, unit: 'Caja', observations: '' }],
+  },
+  {
+    id: 8,
+    code: 'SOL-2024-0008',
+    date: '2024-05-20T13:25:00',
+    area: 'Farmacia',
+    requestedBy: 'Sofia Herrera',
+    status: 'Rechazada',
+    priority: 'Media',
+    observations: 'Producto disponible en stock local del area.',
+    createdAt: '2024-05-20T13:25:00',
+    approvedAt: '2024-05-20T14:01:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Paracetamol 500mg', productCode: 'PROD-001', quantity: 10, unit: 'Caja', observations: '' }],
+  },
+  {
+    id: 9,
+    code: 'SOL-2024-0009',
+    date: '2024-05-19T07:40:00',
+    area: 'Hospitalizacion',
+    requestedBy: 'Elena Vargas',
+    status: 'Pendiente',
+    priority: 'Alta',
+    observations: 'Aumento temporal de camas habilitadas.',
+    createdAt: '2024-05-19T07:40:00',
+    approvedAt: '',
+    approvedBy: '',
+    products: [{ product: 'Guantes de Nitrilo Talla M', productCode: 'PROD-004', quantity: 22, unit: 'Caja', observations: '' }],
+  },
+  {
+    id: 10,
+    code: 'SOL-2024-0010',
+    date: '2024-05-18T10:05:00',
+    area: 'Cirugia',
+    requestedBy: 'Pedro Leon',
+    status: 'Aprobada',
+    priority: 'Alta',
+    observations: 'Jornada quirurgica.',
+    createdAt: '2024-05-18T10:05:00',
+    approvedAt: '2024-05-18T10:40:00',
+    approvedBy: 'Admin SAGI',
+    products: [
+      { product: 'Jeringa 5ml', productCode: 'PROD-005', quantity: 12, unit: 'Caja', observations: '' },
+      { product: 'Alcohol antiseptico 70%', productCode: 'PROD-006', quantity: 20, unit: 'Unidad', observations: '' },
+    ],
+  },
+  {
+    id: 11,
+    code: 'SOL-2024-0011',
+    date: '2024-05-17T16:55:00',
+    area: 'Laboratorio Clinico',
+    requestedBy: 'Camilo Ruiz',
+    status: 'Pendiente',
+    priority: 'Media',
+    observations: 'Turno extendido.',
+    createdAt: '2024-05-17T16:55:00',
+    approvedAt: '',
+    approvedBy: '',
+    products: [{ product: 'Alcohol antiseptico 70%', productCode: 'PROD-006', quantity: 8, unit: 'Unidad', observations: '' }],
+  },
+  {
+    id: 12,
+    code: 'SOL-2024-0012',
+    date: '2024-05-16T12:15:00',
+    area: 'Consulta Externa',
+    requestedBy: 'Paula Mejia',
+    status: 'Aprobada',
+    priority: 'Baja',
+    observations: 'Reposicion rutinaria.',
+    createdAt: '2024-05-16T12:15:00',
+    approvedAt: '2024-05-16T13:02:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Omeprazol 20mg', productCode: 'PROD-002', quantity: 3, unit: 'Caja', observations: '' }],
+  },
+  {
+    id: 13,
+    code: 'SOL-2024-0013',
+    date: '2024-05-15T09:05:00',
+    area: 'Urgencias',
+    requestedBy: 'Monica Silva',
+    status: 'Rechazada',
+    priority: 'Alta',
+    observations: 'Solicitud duplicada.',
+    createdAt: '2024-05-15T09:05:00',
+    approvedAt: '2024-05-15T09:40:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Ibuprofeno 400mg', productCode: 'PROD-003', quantity: 14, unit: 'Caja', observations: '' }],
+  },
+  {
+    id: 14,
+    code: 'SOL-2024-0014',
+    date: '2024-05-14T14:35:00',
+    area: 'Farmacia',
+    requestedBy: 'Daniela Cano',
+    status: 'Pendiente',
+    priority: 'Media',
+    observations: 'Inventario preventivo.',
+    createdAt: '2024-05-14T14:35:00',
+    approvedAt: '',
+    approvedBy: '',
+    products: [{ product: 'Paracetamol 500mg', productCode: 'PROD-001', quantity: 16, unit: 'Caja', observations: '' }],
+  },
+  {
+    id: 15,
+    code: 'SOL-2024-0015',
+    date: '2024-05-13T08:10:00',
+    area: 'Hospitalizacion',
+    requestedBy: 'Oscar Marin',
+    status: 'Aprobada',
+    priority: 'Media',
+    observations: 'Reposicion aprobada.',
+    createdAt: '2024-05-13T08:10:00',
+    approvedAt: '2024-05-13T09:00:00',
+    approvedBy: 'Admin SAGI',
+    products: [{ product: 'Guantes de Nitrilo Talla M', productCode: 'PROD-004', quantity: 10, unit: 'Caja', observations: '' }],
+  },
+];
+
+export const requestAreas = ['Todas las areas', 'Urgencias', 'Farmacia', 'Hospitalizacion', 'Cirugia', 'Laboratorio Clinico', 'Consulta Externa'];
+export const requestStatuses = ['Todos los estados', 'Pendiente', 'Aprobada', 'Rechazada'];
+export const priorityRank = { Baja: 1, Media: 2, Alta: 3, Critica: 4 };
+export const requestsStorageKey = 'sagi_requests';
