+++ frontend/src/pages/auth/RegisterPage.jsx
@@ -1,0 +1,10 @@
+export default function RegisterPage() {
+  return (
+    <main className="auth-shell">
+      <section className="auth-card register-card">
+        <div className="auth-heading"><h1>Registro de usuario</h1><p>Modulo de registro disponible para integracion.</p></div>
+        <a className="auth-primary login-link-button" href="/login">Inicia sesion</a>
+      </section>
+    </main>
+  );
+}
