+++ frontend/src/pages/dashboard/DashboardPage.jsx
@@ -1,0 +1,8 @@
+export default function DashboardPage() {
+  return (
+    <section className="content-panel">
+      <header className="page-header"><div><h1>Dashboard</h1><p>Resumen general del sistema</p></div></header>
+      <section className="inventory-card placeholder-card">Modulo en preparacion.</section>
+    </section>
+  );
+}
