+++ frontend/src/pages/lots/LotsPage.jsx
@@ -1,0 +1,8 @@
+export default function LotsPage() {
+  return (
+    <section className="content-panel">
+      <header className="page-header"><div><h1>Lotes y vencimientos</h1><p>Gestion de lotes y fechas de vencimiento</p></div></header>
+      <section className="inventory-card placeholder-card">Modulo en preparacion.</section>
+    </section>
+  );
+}
