+++ frontend/src/pages/movements/MovementsPage.jsx
@@ -1,0 +1,8 @@
+export default function MovementsPage() {
+  return (
+    <section className="content-panel">
+      <header className="page-header"><div><h1>Movimientos</h1><p>Consulta y registro de movimientos de inventario</p></div></header>
+      <section className="inventory-card placeholder-card">Modulo en preparacion.</section>
+    </section>
+  );
+}
