+++ frontend/src/pages/reports/ReportsPage.jsx
@@ -1,0 +1,8 @@
+export default function ReportsPage() {
+  return (
+    <section className="content-panel">
+      <header className="page-header"><div><h1>Reportes</h1><p>Indicadores y reportes del inventario</p></div></header>
+      <section className="inventory-card placeholder-card">Modulo en preparacion.</section>
+    </section>
+  );
+}
