+++ frontend/src/pages/settings/SettingsPage.jsx
@@ -1,0 +1,8 @@
+export default function SettingsPage() {
+  return (
+    <section className="content-panel">
+      <header className="page-header"><div><h1>Configuracion</h1><p>Parametros generales de SAGI</p></div></header>
+      <section className="inventory-card placeholder-card">Modulo en preparacion.</section>
+    </section>
+  );
+}
