+++ frontend/src/pages/users/UsersPage.jsx
@@ -1,0 +1,8 @@
+export default function UsersPage() {
+  return (
+    <section className="content-panel">
+      <header className="page-header"><div><h1>Usuarios</h1><p>Gestion de usuarios del sistema</p></div></header>
+      <section className="inventory-card placeholder-card">Modulo en preparacion.</section>
+    </section>
+  );
+}
