+++ frontend/src/utils/format.js
@@ -1,0 +1,8 @@
+export const formatRequestDate = (date) =>
+  new Intl.DateTimeFormat('es-CO', {
+    day: '2-digit',
+    month: '2-digit',
+    year: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit',
+  }).format(new Date(date));
