+++ frontend/src/utils/navigation.js
@@ -1,0 +1,4 @@
+export function navigateTo(path) {
+  window.history.pushState({}, '', path);
+  window.dispatchEvent(new PopStateEvent('popstate'));
+}
